backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.5.2
python-dotenv==1.0.0
openai==1.12.0
httpx==0.26.0
pytest==7.4.3
black==23.12.0 